add routes for creating new food items, WIP populating list
dquote> co-authored-by: Raed Alkhanbashi <raed.alkhanbashi@gmail.com>

+++ PackerTracker/Models/Gear.cs
@@ -3,8 +3,8 @@ namespace PackerTracker.Models
   public abstract class Gear
   {
     public string Name { get; set; }
-    public int Weight { get; set; }
-    public int Price { get; set; }
+    public float Weight { get; set; }
+    public float Price { get; set; }
     public bool PurchaseStatus { get; set; }
     public bool PackedStatus { get; set; }
